Store the name of the trace file in the db
At the moment only mirrors that don't list alias(es) are added, since
there's no reliable way to determine the right name otherwise.
It is not even possible to add those that only have one alias. The
mirror entries are stored in a hash map, we are therefore unable to
know if only one alias existed on the mirrors list.