When requesting a more detailed check, do not abort on bad traces
For example, when updating with ./update.sh all the available checks
are run. If a mirror fails because of a bad master or site trace, the
extra checks would be skipped. In such case, it could be possible for a
mirror to be re-enabled even though it wouldn't have been re-enabled if
all the checks had been run in the first place.