Store ids of mirrors that can be considered for the request
The domain name and path to the base directory of the mirror was being
stored instead. Not only this required some operations for mirrors that
could later be discarded but it also meant hashes and other structures
required a bit more space than what was needed.

This change also allows other parts of the code to query the mirror's
properties. They can only be accessed by mirror id.