Redirect user to mirrors that support ranges, when ranges are requested
Putting such a hard block in place might be detrimental. This feature
needs some experimentation and other changes.

Example scenarios:

* Positive:
A user is downloading a file but notices she missed "something" and
cancels the transfer. When attempting to resume the download, if she is
redirected to a mirror that doesn't support ranges she will have to
download the file including the parts she had already downloaded.

* Negative:
A user is downloading a file too slow and might cancel the transfer in
order to try to have better luck with another mirror. If the slow mirror
is the only one in the set that supports ranges, the user won't be sent
to any other and will have to wait until the download finally finishes.

Conflicts:

	redir.pl