Disable mirrors based on available master stamp per continent
The per-region "mirror subsets" feature.

There are some regions where it takes far longer to receive the
updates. With this change, if all the mirrors in a given continent are
behind the latest master stamp, they will not be disabled.

It needs additional work to take into consideration the following:
* There should be a limit as to how old the per-region stamp can be
allowed to be, compared to the master stamp.
* redir.pl is not yet aware of the subsets and may combine mirrors from
different subsets when doing a "nearby-continent" selection.