Reduce precision of math operations
When calculating the standard deviation it could be possible for the
variance to result in a negative number due to float precision issues.
Perl would then die when trying to calculate the sqrt of the variance.

This was seen on the live instance at http.debian.net. Thanks to Simon
Paillard for reporting it.