Store the id of the destination mirror in the peers db, instead of the AS
After some testing and discussing it with Marco d'Itri, the peers db now
stores the id of the destination mirror(s). This allows an AS to be
mapped only to a given mirror, instead of whatever mirror might be in
another AS.

Whenever an ASN is specified instead of a mirror's name, the script will
look up all the mirrors in the given AS and store their ids.