Use the 'Revision' field of trace files, if it exists
Non-ftpsync mirrors may also provide certain required features that
ftpsync mirrors are known to provide. The new 'Revision' allows such
non-ftpsync mirrors to detail what is and, by its omission, what is not
supported.

This is related and closes #14

The new field is described in the following email, but it might be
revised and published elsewhere at a later time (at www.debian.org, for
instance):
https://lists.debian.org/debian-mirrors/2012/07/msg00004.html