Correctly handle some IPv6 tunnel addresses
The code was overly strict and required all four characters in each
block, not taking into account the "shortened" versions where 0s are
omitted.

The code is now overly permissive, allowing addresses such as 0.x.x.x
for 6to4 tunnels, but those invalid addresses: a) shouldn't be seen,
b) would be caught at a later point in the script.