Enable archs-trace check by default
Let's assume for a moment that if the trace file says an architecture
is included, it indeed is. This avoids some tricks that would require
the normal arch-check to be run before re-enabling an architecture that
was disabled via the trace file in the first place.

This assumption can be wrong in many cases. If any of them is ever
seen, the code could be modified so that an architecture is never
re-enabled automatically. I hope that won't be necessary...

Closes #41